[Calibration] Add the resizing option to the calibration code.
It's possible that the calibration data was extracted from an image
of a different resolution (normally higher) than the video flow
resolution. Thus these calibration data needs to be scaled. For this
an imageSize parameter has to be added both to the calibration file
(this value keeps the size of the image used to produce the
calibration data), and to the configuration file (specifies the
target video resolution).