[Grabber] Resize the input video stream to the requested size.
The target image size is specified in the configuration file and
is used by the calibration routine. It seems logical to use this
values in the grabber to resize input video flow to this size
thus insuring  that the futher calculations are correct. To read
and to store target image size from the configuration file a new
class GrabberSettings is added to the grabber library.