[Calibration] Add new hub library.
The input video stream is treated by the tracking setup in one
connected data flow, but if we need to show to an user the video
then a kind of hub is needed to send the original data out of the
tracking setup. Such a functionality was implemented in the hub
library. Upon request (passed as a flag in the constructor), the
tracking setup creates a queue hub which is used to pass the images
to the tracker but also to a supplimentary queue that can be used
from the outside.