Add the ViewerSettings class.
Since we have the information on the frame size for every setup in
the configuration file it makes sense to use it to adjust the size
of the viewer's view. This information is loaded and stored in the
ViewerSettings class, and it is provided to the constructor of the
 ViewerWidget class by the ViewerHandler; the latter is also given
the setup type to know which data is get from the settings.