Add specific settings widgets for every tracking routine.
To set the settings of specific routines corresponsing widgets are added. They
are created by TrackingUiFactory class. In future it will be possible to save
the settings to the configurations file, at the moment all the changes are
forgot once the application is closed. To prevent from modifying the settings
while they are used by the routine they are protected in the routine clases by a
mutex.