Fine-tune the coordinates conversion.
The pixel position of the object whose height is different from the height of
the calibration markers will be different even if the (x,y) position is the
same (due to the perspective). Thus the height has to be taken into account.
At the moment we added one value to the calibration file that represents an
average altitude of the agents with respect to the calibration data. This value
is used every time an image position from this setup if converted to a world
position.