Send tracking results via zmq in main camera frame coordinates.
The original code sent the tracking results in world coordinates, while CATS
that receives these data expects image coordinates. To fix this
TrackingDataManager class got an access to all coordinates conversion classes,
it uses the main camera calibration to generate a agent list in image
coordinates, and sends the results with notifyAgentDataImageMerged signal.