Add the navigation functionality to the robot control.
It excecutes the control target in its step() method. The result is the motors
commands that are sent to robots. Two motion patterns are supported: basic PID
and a fish motion pattern. The navigation object is owned by the robot. Parame-
ters of these motion patterns are read from configuration file and are stored
in the RobotControlSettings.