Add the control map concept.
Two classes are added : ControlMap that covers the whole experimental area and
ControlArea that represents a part of the map that is defined by the same
control behavior. The control map for every given position of the experimental
setup provides the control mode and motion pattern to excecute in this position.
Every robot has it's own control map that it consults on every step.