Add command frequency divider for the motion pattern.
The more often fish pattern is sent to the robot the more chaotically it moves
that corresponds to a stressed fish. Thus we added a frequency divider with
respect to the fixed control loop frequency, that defines that the fish pattern
command is to be sent every M-th step. The value is read from the configuration
file.

This behavior might be in future extended to the PID controller.