Send only one timestamp with the tracking results.
Before every tracked agent had its timestamp. But since we always used only one
timestamp for the whole batch of results, now we attach the timestamp to all the
data. The advantage is that like this even if the tracking doesn't find any
agents we can send an empty list of tracked agents with the correct timestamp
(corresponding to the frame), and thus we don't block the tracking result's
matching and the visualization.