Apply a quick hack to the lost fish in the log problem.
The fish tracking is not reliable, the danger is that ids swap and that when a
robot is invisible by the tracking from below it's tracked as a fish and thus
"steals" someone's id, the corresponding fish gets a bigger id and is not seen
anymore by the trajectory writer. We apply a quick and dirty hack that makes
the writer to forget all the mapping at every step and thus localize the above-
mentioned problem only to frames when the robot was not properly tracked.