Do major refactoring of the grid based methods.
(1) The grid now is bigger than the setup to include it completely. Like this
we quarantee that the grid includes all the walls that are used by the
potential field obstacle avoidance.
(2) The grid based methods now first use the grid to check if a point is in the
free area, and only if it's not possible they consult the setup map. This
approach is considerably faster.
(3) Dijkstra path planning is consulting the grid to build its graph instead of
using the setup map direclty as before.