Add a tracking routine for two-color tags.
A very simple idea of one tag group in the front of the robot and the second in
the back. The orientation is defined as a line between  centers of mass of two
groups of tags and the center as a point on this line shifted by a predefined
coefficient. It supports only one agent at the moment.