Update the initiation controller for the leader robot.
It follows the request of Yohann:
(1) The robot follow the fish model (not restricted to specific room) and does
the fish-motion pattern.

(2) IF the robot finishes in the Other-Room BUT the most of the fish is not
with him then it continue (1).

(3) IF the robot finishes in the Other-Room AND the most of the fish is with
 him AND there are fish close to the robot THEN the robot goes to the center
 of the  Target-Room, first for one second in the fish-motion, and then in
PID. Once it arrives to the Target-Room it switches again to the  fish-motion,
 but still makes circles around the center of the room.

(4) When the robot starts going to the target room a timer is started, on its
 timeout the robot checks if the fish follow, IF they don’t follow THEN the
robot returns back to the Other-Room in PID, once on the target it switches
to (1).

(5) IF the robot is in the Target-Room AND the fish follow it THEN it waits
until there is at least one fish in the proximity, in this case it switches
to (1).